+++ src/main/java/lotto/validation/BonusNumberValidator.java
@@ -1,0 +1,2 @@
+package lotto.validation;public class BonusNumberValidator {
+}
